refactor: workspace 관련 layout 폴더링 리팩토링
