+++ app/(workspace)/@sidebar/default.tsx
@@ -1,4 +1,4 @@
-import { Sidebar } from '@/widget/layout/workspace';
+import { Sidebar } from '@/widget/workspace';
 
 export default function page() {
   return <Sidebar />;
