--- src/widget/layout/workspace/index.ts
@@ -1,2 +1,0 @@
-export { Header } from './Header';
-export { Sidebar } from './Sidebar';
