+++ src/widget/workspace/index.ts
@@ -1,0 +1,2 @@
+export { Header } from './layout/Header';
+export { Sidebar } from './layout/Sidebar';
