+++ src/widget/workspace/layout/Sidebar.tsx
